requirements.txt
aniso8601==7.0.0
asgiref==3.3.0
Django==3.0.5
django-cors-headers==3.5.0
django-environ==0.4.5
django-graphql-jwt==0.3.1
django-restql==0.10.2
djangorestframework==3.12.1
djongo==1.3.3
graphene==2.1.8
graphene-django==2.13.0
graphene-mongo==0.2.13
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
iso8601==0.1.13
mongoengine==0.20.0
Pillow==8.0.1
promise==2.3
PyJWT==1.7.1
pymongo==3.11.0
pyPEG2==2.15.2
pytz==2020.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.2.4
Unidecode==1.1.1
